backend/requirements.txt
flask==2.0.0
werkzeug==2.0.0
mysql-connector-python==8.0.26
flask-cors==3.0.10
python-dotenv==0.19.0
